feat: Add projection + partial updates
This commit adds the ability for a contexts to load
only a set of fields from the storage.
This can save a lot of time an memory especially
when some fields contain large data.

The problem with implementing projection was that
when you saved the context, all fields that were
not loaded were lost. This commit fixes this by
saving only the fields that were loaded via partial
updates.